moved database typedefs to external file
